requirements.txt
PyYAML==6.0.2
pytest==8.3.5
flake8==7.1.1
